+++ app/src/androidTest/java/it/unimib/brewday/BottomNavigationTest.java
@@ -40,6 +40,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import it.unimib.brewday.ui.MainActivity;
+
 
 @RunWith(AndroidJUnit4.class)
 public class BottomNavigationTest {
