+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="it.unimib.brewday.MainActivity">
+    tools:context="it.unimib.brewday.ui.MainActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/top_appbar"
